integrate-LLM-in-Application/healthcare-telemedicine/requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
moto>=4.2.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
